backend/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0
pytz==2025.2
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
wheel==0.45.1
